+++ src/GafferSceneUI/RotateTool.cpp
@@ -272,7 +272,7 @@ bool RotateTool::buttonPress( const GafferUI::ButtonEvent &event )
 	//
 	// We always return true to prevent the SelectTool defaults.
 
-	if( selection().size() == 0 )
+	if( !selectionEditable() )
 	{
 		return true;
 	}